Remove redundant if statement
This PR removes a redundant if statement, that was never hit by tests.

The reason for this refactoring is that once `value` attribute exists,
we can right away compare its value with `null` string.
If it does match, then it is indeed a `SendNull` pattern.
If `value` contains anything but `null` string, including other types,
then it does not match `SendNull` pattern.

Closes #709